requirements.txt
streamlit-lottie
numpy
pandas